Adding Events
Click on a date block and see what you can do!!
A new modal will appear that allows you to add an event to the specified date and
store it in localStrage.
This functionality is a must in a calendar. A user can now also customize
the color of the event they want to store.

Future updates to this functionality will be to implement repeating
events, times & notifications for each event