Calendar With Repeats, Reminders & More
What use is a Calendar without a system to rekind you?
or repeat tasks that matter and anually, monthly or
weekly come about? No good is the amswer.

Realizing the Modal component will end up bwing quite large
i satisfyingly accept that and begin writting the
functionality to make this Calendar part of a
next generation tool.