User Context
A new user context was created to check weather or not a client is logged in.

This application will still persist and work correctly weather or not you are logged in. Events will be stored in local storage
if a client is not logged in to maintain functionality.

A new backend was also created to handle authentication and authorization via express and JWT. Prisma client will be used
to query the database as a usefull ORM