Auth
JWT is now implemented into the google login route and sent down to the user to fetch all events tied to that user.

A way to keep the user persistent and in conjunction with the token persistence will be the next challenge. This
will allow for a better user experience and user flow.