Time Allotted
Now events are displayed in the modal as a day view section where the start and end times coordinate with the static 24 hour period of the day.
This makes for a great user experience as it allows the user to see a month view and a day view right off the bat.