Notifications!??
Notifications Via Service Worker are now implimented into the application via express and web push on the backend.

Attempting a new way to store notifications on the client side with localStorage has so far failed.
Considering adding the notifications to a new database table of its own to store notifications of the user and allow persistent data
throughout different devices