Fixing Errors
Admittedly injectManifest config for VitePWA application turns out
to be a bit above my paygrade for now.

an automated service worker will now handle the caching.

Keep in mind that this is temporary. In order for me to clean up the code & start fresh I am using this technique.
Push notifications will still be a major part of this application amd the knowledge on how to integrate them with VitePWA & workbox will be gained. For now, only inapp notifications will be accepted.