Did I finally do it??
That is the question.. Did I finally and successfully figure out how to add the default workbox precaching and updating strategies to this
fantastic application while also adding in the functionality to subscribe and listen for true real push notifications? I do not know.

Here goes a production (possibly read) code base to test it out.