Event In Context && WeekView
Now Event component logic is placed within the context to help make it more reusable.

WeekView was started by conpleting a useEffect hook that will render date objects for each day of the week that
you are currently in starting with Sunday and going to Saturday. These date object are then passed down to the WeekView componenet to be used and rendered