Wow
Massive update to frontend code, and a few backend typos.

As a result of failure to update the applications storage and active states for the user and authentication token after
subscribing to notifications caused a massive bug where a user no longer could recieve notifications or properly subscribe to them.