Masonry Grid && Notification Badge
Now the application has a new notification badge made for that specific purpose.
The event now has the area for a google maps the user can view and interact with along with open
the google maps directions in the app or on the internet if the app is not present.

The Masonry Grid layout now has a useRef and useEffect that brings the user to the closest period of time events are relevent