Small Changes, Big Results
Fixing display standalone issues with pwa application. Creating a static manifest.json file
and forgetting to update the display on the static file was causing the failure to update the
applications display.

Now a user will create an event and handle attachments seperately form the event. This allows
the user to create attachments and upload large files to the server while keeping
full functionality to the application open and not interupting the user experience.