Menu UI/UX
Changing the header when the menu is open and updating the view and add buttons on the bottom
right and left corners to accomidate what context the user is in.

Connections state is now being implemented so the user is capable of creating connections
and making friends on the application. Friend requests will be sent via email and
notifications in app