Performance & UI/UX
Updates to month views rendering logic created for an increase in speed and performace by 25%.
This also allows me to create indexing on events within a day to provide styling and margin on the tops and bottoms
of each event within the day after setting them to position fixed for multiple day events.

Style and animation updates were introduced to give a user a better feel for speed of the application.