Updating Prisma CLient
Accepting a more structured kanban board schema and keeping information organized.

For now, instead of creating more tables to create solid relationships with each piece of data in the boardItems
schema, Json? was set to keep things more dynamic for now until a solid schema is established and functionality is needed.