Serious Backend/Frontend Refactoring
Using prisma includes will decrease significantly the amount of queries being made to the DB.

Because user friends data is special and sensitive only information on the server is passed around. No user has access to other user information
other than email and permissions. In this instance the friend request/sent/recieved and friends in general were not included in the prisma
includes.

Notifications are also a special bread in this application and were left out of the prisma includes query.

The structure of how a user logs in or signs up is now slightly different.

Only a token is sent back on initial signup/login attempts and user data is then seperatly fetched. This way the user is only being
set once in state and no data is reaching the client unless a valid token is successfully sent back to the server.