Kanban Creation
Now a user can create a kanban with folders on his profile.
Both frontend and and server logic were wrutteb for a successful and safe way
with error handling.

Folders are created seperately from the kanban board on the server by first creating a kanban and
then grabbing the kanban id so accosiate the folders with it. the folders are mapped through and
created in the database with a reference to the kanban board

A few Minor changes to the code base were implemented as well. Adding a key to the sticky and kaban data being mapped through,
as well as building the simple UI for displaying a kanban board